[chore][pkg/stanza] Localize one-off test utility code
